pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ---------------------------------------------------------------------------
# Project metadata  (PEP 621)
# ---------------------------------------------------------------------------
[project]
name = "afterimage"
version = "1.0.0"
description = "Optical air-gap data exfiltration via QR stream with LT Fountain Codes"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
authors = [
    { name = "AFTERIMAGE Contributors" },
]
keywords = [
    "air-gap",
    "exfiltration",
    "qr-code",
    "fountain-codes",
    "cryptography",
    "optical",
    "steganography",
    "security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Multimedia :: Video :: Capture",
    "Topic :: Communications",
    "Typing :: Typed",
]
requires-python = ">=3.10"

# Core runtime dependencies
dependencies = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "cryptography>=41.0",
    # QR generation — segno is faster; qrcode is the fallback
    "segno>=1.6",
    "Pillow>=10.0",
]

[project.optional-dependencies]
# Faster QR scanning (requires libzbar system library)
pyzbar = ["pyzbar>=0.1.9"]

# Development / testing
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-Pillow",
]

# Documentation
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=3.0",
]

[project.urls]
Homepage       = "https://github.com/nzengi/AfterImage"
Documentation  = "https://github.com/nzengi/AfterImage#readme"
Repository     = "https://github.com/nzengi/AfterImage"
"Bug Tracker"  = "https://github.com/nzengi/AfterImage/issues"
Changelog      = "https://github.com/nzengi/AfterImage/blob/main/CHANGELOG.md"

[project.scripts]
afterimage = "afterimage.cli:main"

# ---------------------------------------------------------------------------
# Tool: setuptools
# ---------------------------------------------------------------------------
[tool.setuptools.packages.find]
where = ["."]
include = ["afterimage*"]

[tool.setuptools.package-data]
afterimage = ["py.typed"]

# ---------------------------------------------------------------------------
# Tool: pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short --strict-markers"
markers   = [
    "slow: marks tests as slow (deselect with -m 'not slow')",
    "camera: marks tests that require a physical camera",
]

# ---------------------------------------------------------------------------
# Tool: coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source   = ["afterimage"]
omit     = ["afterimage/__main__.py", "afterimage/cli.py"]
branch   = true

[tool.coverage.report]
show_missing  = true
skip_covered  = false
fail_under    = 80

# ---------------------------------------------------------------------------
# Tool: ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length    = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "N",   # pep8-naming
]
ignore = [
    "S101",  # assert statements (used in tests)
    "S311",  # standard pseudo-random (numpy RNG is fine here)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "N"]

# ---------------------------------------------------------------------------
# Tool: mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version      = "3.10"
strict              = true
ignore_missing_imports = true